Subclass refactor.
